pyproject.toml
[project]
name = "rpi-led-matrix-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">= 3.10, <3.12"
dependencies = [
  "pillow>=10.4.0",
  "websockets>=14.1",
]
